requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
SQLAlchemy==2.0.23

# Password Hashing
bcrypt==4.1.2

# NLP & AI
spacy==3.7.2

# File Handling
python-magic==0.4.27

# CSV Export
pandas==2.1.4

# Text Similarity
scikit-learn==1.3.2

# Environment Variables
python-dotenv==1.0.0

# OpenAI
openai==1.12.0
